requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
web3>=6.0.0
